refactor backtester
