+++ wayfinder_paths/core/backtesting/constants.py
@@ -1,0 +1,11 @@
+"""Constants for backtesting framework."""
+
+DEFAULT_MAINTENANCE_MARGINS = {
+    "HYPE/USDC:USDC": 1 / 20.0,
+    "ASTER/USDC:USDC": 1 / 16.0,
+    "DYDX/USDC:USDC": 1 / 20.0,
+    "GMX/USDC:USDC": 1 / 20.0,
+    "APEX/USDC:USDC": 1 / 20.0,
+    "AVNT/USDC:USDC": 1 / 10.0,
+    "BTC/USDC:USDC": 1 / 100.0,
+}
